Move generic I/O functions to plugins/module_utils/io.py.
