+++ plugins/modules/ecs_certificate.py
@@ -528,8 +528,8 @@ from distutils.version import LooseVersion
 from ansible.module_utils.basic import AnsibleModule, missing_required_lib
 from ansible.module_utils._text import to_native, to_bytes
 
-from ansible_collections.community.crypto.plugins.module_utils.crypto.basic import (
-    write_file
+from ansible_collections.community.crypto.plugins.module_utils.io import (
+    write_file,
 )
 
 from ansible_collections.community.crypto.plugins.module_utils.crypto.support import (
