+++ plugins/modules/openssl_pkcs12.py
@@ -187,11 +187,14 @@ import traceback
 from ansible.module_utils.basic import AnsibleModule, missing_required_lib
 from ansible.module_utils._text import to_bytes, to_native
 
+from ansible_collections.community.crypto.plugins.module_utils.io import (
+    load_file_if_exists,
+    write_file,
+)
+
 from ansible_collections.community.crypto.plugins.module_utils.crypto.basic import (
     OpenSSLObjectError,
     OpenSSLBadPassphraseError,
-    load_file_if_exists,
-    write_file,
 )
 
 from ansible_collections.community.crypto.plugins.module_utils.crypto.support import (
