+++ plugins/modules/openssl_privatekey.py
@@ -282,6 +282,11 @@ from distutils.version import LooseVersion
 from ansible.module_utils.basic import AnsibleModule, missing_required_lib
 from ansible.module_utils._text import to_native, to_bytes
 
+from ansible_collections.community.crypto.plugins.module_utils.io import (
+    load_file_if_exists,
+    write_file,
+)
+
 from ansible_collections.community.crypto.plugins.module_utils.crypto.basic import (
     CRYPTOGRAPHY_HAS_X25519,
     CRYPTOGRAPHY_HAS_X25519_FULL,
@@ -290,8 +295,6 @@ from ansible_collections.community.crypto.plugins.module_utils.crypto.basic impo
     CRYPTOGRAPHY_HAS_ED448,
     OpenSSLObjectError,
     OpenSSLBadPassphraseError,
-    write_file,
-    load_file_if_exists,
 )
 
 from ansible_collections.community.crypto.plugins.module_utils.crypto.support import (
