+++ plugins/modules/x509_crl.py
@@ -354,10 +354,13 @@ from distutils.version import LooseVersion
 from ansible.module_utils.basic import AnsibleModule, missing_required_lib
 from ansible.module_utils._text import to_native, to_text
 
+from ansible_collections.community.crypto.plugins.module_utils.io import (
+    write_file,
+)
+
 from ansible_collections.community.crypto.plugins.module_utils.crypto.basic import (
     OpenSSLObjectError,
     OpenSSLBadPassphraseError,
-    write_file,
 )
 
 from ansible_collections.community.crypto.plugins.module_utils.crypto.support import (
